moved animation config to notification component
