+++ app/shell/Notifications/Notifications.tsx
@@ -1,4 +1,4 @@
-import { AnimatePresence, motion, Transition, Variants } from 'framer-motion';
+import { AnimatePresence } from 'framer-motion';
 import React from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import Notification from '~/components/Notification';
@@ -11,21 +11,6 @@ import { RootState } from '~/store/store';
 
 import styles from './Notifications.module.scss';
 
-const transition: Transition = {
-    ease: 'easeIn',
-    duration: 0.2,
-};
-
-const transitionVariants: Variants = {
-    initial: { opacity: 0, transform: 'translate(100%, 0)' },
-    enter: { opacity: 1, transform: 'translate(0, 0)' },
-    exit: {
-        opacity: 0,
-        transform: 'translate(0, -100%)',
-        transition: { ease: 'easeOut' },
-    },
-};
-
 export default function Notifications() {
     const { notifications } = useSelector((s: RootState) => s.notification);
     const dispatch = useDispatch();
@@ -37,31 +22,24 @@ export default function Notifications() {
     return (
         <Portal className={styles.root}>
             <AnimatePresence>
-                {notifications.map((n) => (
-                    <motion.div
-                        key={n.id}
-                        transition={transition}
-                        variants={transitionVariants}
-                        initial="initial"
-                        animate="enter"
-                        exit="exit"
-                    >
-                        {n.level === NotificationLevel.Update ? (
-                            // TODO: Replace with update notification
-                            <Notification
-                                level={n.level}
-                                onCloseClick={() => handleClose(n.id)}
-                            />
-                        ) : (
-                            <Notification
-                                level={n.level}
-                                onCloseClick={() => handleClose(n.id)}
-                            >
-                                {n.message}
-                            </Notification>
-                        )}
-                    </motion.div>
-                ))}
+                {notifications.map((n) =>
+                    n.level === NotificationLevel.Update ? (
+                        // TODO: Replace with update notification
+                        <Notification
+                            key={n.id}
+                            level={n.level}
+                            onCloseClick={() => handleClose(n.id)}
+                        />
+                    ) : (
+                        <Notification
+                            key={n.id}
+                            level={n.level}
+                            onCloseClick={() => handleClose(n.id)}
+                        >
+                            {n.message}
+                        </Notification>
+                    )
+                )}
             </AnimatePresence>
         </Portal>
     );
